requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
eventlet==0.33.3
numpy<2.0.0
pandas==2.1.0
pdfkit==1.0.0
Werkzeug==2.3.7
python-engineio==4.7.1
python-socketio==5.9.0
gunicorn==21.2.0
openpyxl==3.1.2
xlsxwriter==3.1.9
psycopg2-binary==2.9.9
